cambiado int a double

+++ freeglut project 2D/pelota.h
@@ -11,7 +11,7 @@
 #define PELOTA_H
 
 const int SEGMENTOS_CIRCULO=20;
-const int M_PI= 3.1415926535897932384626433832795;
+const double M_PI= 3.1415926535897932384626433832795;
 
 #include "PV2D.h"
 #include "lapiz.h"
